requirements.txt
numpy>=1.26
pandas>=2.2
torch>=2.5
transformers>=4.45
accelerate>=1.3
tqdm>=4.67
openai>=1.6
python-dotenv>=1.0
pyyaml>=6.0
huggingface_hub>=0.24
pydantic>=2.0
pydantic-ai>=0.1
